agent/requirements.txt
openai-agents>=0.12.0
python-dotenv>=1.0.0
